[#163518700] rename from shareValidtor to article validator
